// src/lib/enums.ts
export enum PlanType {
  Monthly = "Monthly",
  Quarterly = "Quarterly",
  Half_Yearly = "Half_Yearly",
  Yearly = "Yearly",
}

export enum SubscriptionCategory {
  Music = "Music",
  OTT = "OTT",
  Fitness = "Fitness",
  Education = "Education",
  News = "News",
  Gaming = "Gaming",
  CloudStorage = "Cloud Storage",
  Productivity = "Productivity",
  ECommerce = "E-Commerce",
}